move not specific function from utils
